+++ imap_processing/tests/mag/test_mag_l1b.py
@@ -201,3 +201,19 @@ def test_calibrate_vector():
     expected_vector = [4584.1029091, 27238.73161294, -38405.22240195, 0.0]
 
     assert np.allclose(cal_vector, expected_vector, atol=1e-9)
+
+
+def test_l1a_to_l1b(validation_l1a):
+    # Convert l1a input validation packet file to l1b
+    with pytest.raises(ValueError, match="Raw L1A"):
+        mag_l1b(validation_l1a[0], "v000")
+
+    l1b = [mag_l1b(i, "v000") for i in validation_l1a[1:]]
+
+    assert len(l1b) == len(validation_l1a) - 1
+
+    assert l1b[0].attrs["Logical_source"] == "imap_mag_l1b_norm-mago"
+    assert l1b[1].attrs["Logical_source"] == "imap_mag_l1b_norm-magi"
+
+    assert len(l1b[0]["vectors"].data) > 0
+    assert len(l1b[1]["vectors"].data) > 0
